requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
config==0.5.1
coverage==7.6.4
cryptography==43.0.3
defusedxml==0.8.0rc2
Django==5.1.2
django-filter==24.3
django-jazzmin==3.0.1
djangorestframework==3.15.2
idna==3.10
oauthlib==3.2.2
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
python3-openid==3.2.0
requests==2.32.3
requests-oauthlib==2.0.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
typing_extensions==4.12.2
urllib3==2.2.3
